* Added one more set of data to each run through the SSE2 NTLM algorithm.

+++ branches/unstable/newKeyGen/source/KeyGenerator.h
@@ -10,7 +10,7 @@ public:
 	keyGenerator(unsigned long long location, char* charset)
 		: charset(charset), location(location)
 	{
-		charsetLength = strlen(charset);
+		charsetLength = (unsigned int)strlen(charset);
 		integerToKey();
 	}
